+++ lnm/model/MOUs/InstitutionMou.cs
@@ -11,27 +11,16 @@ namespace model.MOUs
     {
         [Key]
         public int ImoMouNo { get; set; }
-
-        public int? ImoMouId { get; set; }
-
+        public float? ImoMouId { get; set; }
         public int? ImoInstitutionId { get; set; }
-
         public int? ImoCreatedBy { get; set; }
-
         public int? ImoUpdatedBy { get; set; }
-
         public DateTime? ImoCreatedDate { get; set; }
-
         public DateTime? ImoUpdatedDate { get; set; }
-
         public float ImoMouVerionNo { get; set; }
-
         public string? ImoMouPath { get; set; }
-
         public string? ImoMouDescription { get; set; }
-
         public string? ImoUploadedMou { get; set; }
-
         public string? ImoDownloadedMou { get; set; }
     }
 }
